GH-4004 Make QueryTypes independend of HTTP and provide it in a reusable way
(cherry picked from commit 934b9c452d39f2b6eba821f04bb38a47b13484f9)
