+++ spring-components/spring-boot-sparql-web/src/main/java/org/eclipse/rdf4j/http/server/readonly/EvaluateResultHttpResponse.java
@@ -1,0 +1,37 @@
+/*******************************************************************************
+ * Copyright (c) 2022 Eclipse RDF4J contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Distribution License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ *******************************************************************************/
+package org.eclipse.rdf4j.http.server.readonly;
+
+import java.io.IOException;
+import java.io.OutputStream;
+
+import javax.servlet.http.HttpServletResponse;
+
+import org.eclipse.rdf4j.http.server.readonly.sparql.EvaluateResult;
+
+/**
+ * Encapsulated the {@link HttpServletResponse}.
+ */
+class EvaluateResultHttpResponse implements EvaluateResult {
+
+	private HttpServletResponse response;
+
+	public EvaluateResultHttpResponse(HttpServletResponse response) {
+		this.response = response;
+	}
+
+	@Override
+	public void setContentType(String contentType) {
+		response.setContentType(contentType);
+	}
+
+	@Override
+	public OutputStream getOutputstream() throws IOException {
+		return response.getOutputStream();
+	}
+}
